Hold personvalg-panel åpent ved satte verdier
